Add fpt

+++ src/isPrime.cpp
@@ -1,22 +1,22 @@
 #include <omp.h>
 #include <math.h>
 #include <iostream>
-bool isPrime(unsigned long n)
+bool isPrime(unsigned long long n)
 {
 	if(n  < 2 || (n != 2 && n%2==0)	) return false;
-	for(unsigned long i = 2; i<sqrt(n); i++)
+	for(unsigned long long i = 2; i<sqrt(n); i++)
 		if(n%i == 0)
 			return false;
 	return true;
 }
 
-bool isPrime_par(unsigned long n)
+bool isPrime_par(unsigned long long n)
 {
 	if(n  < 2 || (n != 2 && n%2==0)	) return false;
-	unsigned long sn = ceil(sqrt(n));	
+	unsigned long long sn = ceil(sqrt(n));	
 	bool ans = true;
 	#pragma omp parallel for
-	for(unsigned long i = 2; i<sn ; i++)
+	for(unsigned long long i = 2; i<sn ; i++)
 	{
 		#pragma omp flush (ans)
 		if(ans)
