+++ src/ss.cpp
@@ -2,16 +2,16 @@
 
 using namespace std;
 
-bool ss(unsigned long n, unsigned int k)
+bool ss(unsigned long long n, unsigned int k)
 {
 	if (n != 2 && n%2 == 0 || n<2) 
 		return false;
 	seed_rand();
 	for(unsigned int i = 0; i < k; i++)
 	{
-		unsigned long a = rand()%(n-1)+1;
-		unsigned long jacobian = (n+jacobi(a,n))%n;
-		unsigned long mod = congruent_mod(a,(n-1)/2,n);
+		unsigned long long a = rand()%(n-1)+1;
+		unsigned long long jacobian = (n+jacobi(a,n))%n;
+		unsigned long long mod = congruent_mod(a,(n-1)/2,n);
 
 		if (!jacobian || mod != jacobian)
 			return false;
@@ -19,7 +19,7 @@ bool ss(unsigned long n, unsigned int k)
 	return true;
 }
 
-bool ss_par(unsigned long n, int k)
+bool ss_par(unsigned long long n, int k)
 {
 	if (n!=2 && n%2 == 0 || n<2) 
 		return false;
@@ -32,9 +32,9 @@ bool ss_par(unsigned long n, int k)
 		#pragma omp flush (ans)
 		if(ans)
 		{
-			unsigned long a = rand()%(n-1)+1;
-			unsigned long jacobian = (n+jacobi(a, n))%n;
-			unsigned long mod = congruent_mod(a,(n-1)/2,n);
+			unsigned long long a = rand()%(n-1)+1;
+			unsigned long long jacobian = (n+jacobi(a, n))%n;
+			unsigned long long mod = congruent_mod(a,(n-1)/2,n);
 
 			if (!jacobian || mod != jacobian)
 			{
